+++ src/main/java/com/esboco_comix/dao/mapper/ResultSetMapper.java
@@ -3,7 +3,7 @@ package com.esboco_comix.dao.mapper;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-public interface ResultSetMapper<Entidade, DTO> {
-    public Entidade mapearEntidade(ResultSet rs) throws SQLException;
-    public DTO mapearDTO(ResultSet rs) throws SQLException;
+public interface ResultSetMapper<E, D> {
+    E mapearEntidade(ResultSet rs) throws SQLException;
+    D mapearDTO(ResultSet rs) throws SQLException;
 }
